preliminary tidying up
