+++ rust/fastsim-core/fastsim-proc-macros/src/lib.rs
@@ -1,9 +1,16 @@
+// modules
 mod imports;
-use crate::imports::*;
+// modules - macros
 mod add_pyo3_api;
 mod approx_eq_derive;
+mod doc_field;
 mod history_vec_derive;
+
+// modules - other
 mod utilities;
+
+// imports
+use crate::imports::*;
 
 /// macro for creating appropriate setters and getters for pyo3 struct attributes
 #[proc_macro_error]
@@ -11,6 +18,12 @@ mod utilities;
 pub fn add_pyo3_api(attr: TokenStream, item: TokenStream) -> TokenStream {
     add_pyo3_api::add_pyo3_api(attr, item)
 }
+
+// #[proc_macro_error]
+// #[proc_macro_attribute]
+// pub fn doc_field(attr: TokenStream, item: TokenStream) -> TokenStream {
+//     doc_field::doc_field(attr, item)
+// }
 
 #[proc_macro_derive(HistoryVec)]
 pub fn history_vec_derive(input: TokenStream) -> TokenStream {
